requirements.txt
streamlit==1.28.0
boto3==1.28.44
Pillow==10.0.0